microservices/requirements.txt
flask==3.0.0
flask-cors==4.0.0
nltk==3.8.1
textblob==0.17.1
librosa==0.10.1
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
